+++ private/steps/5-drivers/step-drivers-addwindowsdriver-firmware.ps1
@@ -17,7 +17,7 @@ function step-drivers-addwindowsdriver-firmware {
         if (-not (Test-Path -Path $LogPath)) {
             New-Item -ItemType Directory -Path $LogPath -Force | Out-Null
         }
-        Add-WindowsDriver -Path "C:\" -Driver "$DriverPath" -Recurse -ForceUnsigned -LogPath "$LogPath\drivers-firmware.log" -ErrorAction SilentlyContinue | Out-Null
+        Add-WindowsDriver -Path "C:\" -Driver "$DriverPath" -Recurse -ForceUnsigned -LogPath "$LogPath\drivers-firmware.log" -ErrorAction SilentlyContinue
     }
     #=================================================
     # End the function
