+++ public/Deploy-OSDCloud.ps1
@@ -12,6 +12,13 @@ function Deploy-OSDCloud {
         $CLI
     )
     #=================================================
+    Write-Host -ForegroundColor DarkCyan "OSDCloud collects analytic data during the deployment process to identify issues, enhance performance, and improve the overall user experience."
+    Write-Host -ForegroundColor DarkCyan "No personally identifiable information (PII) is collected, and all data is anonymized to protect user privacy."
+    Write-Host -ForegroundColor DarkCyan "Collected data includes information about the deployment environment and system configuration."
+    Write-Host -ForegroundColor DarkCyan "By using OSDCloud, you consent to the collection of analytic data as outlined in the privacy policy"
+    Write-Host -ForegroundColor DarkGray "https://github.com/OSDeploy/OSDCloud/blob/main/PRIVACY.md"
+    Write-Host ""
+    #=================================================
     # Initialize OSDCloudWorkflow
     Initialize-OSDCloudDeploy -WorkflowName $WorkflowName
     #=================================================
