+++ workflow/default-back/os-arm64.json
@@ -1,0 +1,67 @@
+{
+  "OperatingSystem.default": "Windows 11 25H2",
+  "OperatingSystem.values": [
+    "Windows 11 25H2",
+    "Windows 11 24H2",
+    "Windows 11 23H2"
+  ],
+  "OSActivation.default": "Retail",
+  "OSActivation.values": ["Retail", "Volume"],
+  "OSEdition.default": "Pro",
+  "OSEditionId.default": "Professional",
+  "OSEdition.values": [
+    {
+      "Edition": "Home",
+      "EditionId": "Core"
+    },
+    {
+      "Edition": "Pro",
+      "EditionId": "Professional"
+    },
+    {
+      "Edition": "Enterprise",
+      "EditionId": "Enterprise"
+    }
+  ],
+  "OSLanguageCode.default": "en-us",
+  "OSLanguageCode.values": [
+    "ar-sa",
+    "bg-bg",
+    "cs-cz",
+    "da-dk",
+    "de-de",
+    "el-gr",
+    "en-gb",
+    "en-us",
+    "es-es",
+    "es-mx",
+    "et-ee",
+    "fi-fi",
+    "fr-ca",
+    "fr-fr",
+    "he-il",
+    "hr-hr",
+    "hu-hu",
+    "it-it",
+    "ja-jp",
+    "ko-kr",
+    "lt-lt",
+    "lv-lv",
+    "nb-no",
+    "nl-nl",
+    "pl-pl",
+    "pt-br",
+    "pt-pt",
+    "ro-ro",
+    "ru-ru",
+    "sk-sk",
+    "sl-si",
+    "sr-latn-rs",
+    "sv-se",
+    "th-th",
+    "tr-tr",
+    "uk-ua",
+    "zh-cn",
+    "zh-tw"
+  ]
+}
