+++ workflow/default-back/tasks/osdcloud.json
@@ -1,0 +1,211 @@
+{
+    "id": "2bc42a9c-3ce1-44a6-847d-593c47d362c5",
+    "name": "OSDCloud",
+    "description": "OSDCloud",
+    "author": "osdcloud@recastsoftware.com",
+    "version": "2026.2.5.1",
+    "amd64": true,
+    "arm64": true,
+    "default": true,
+    "steps": [
+        {
+            "name": "Initialize OSDCloud Workflow",
+            "command": "step-initialize-osdcloudworkflowtask",
+            "parameters": {
+                "WorkflowTaskName": "OSDCloud"
+            },
+            "pause": false,
+            "skip": false,
+            "debug": false,
+            "verbose": false,
+            "testinfullos": true
+        },
+        {
+            "name": "Initialize OSDCloud Logs",
+            "command": "step-initialize-osdcloudlogs",
+            "testinfullos": true
+        },
+        {
+            "name": "Test TargetDisk",
+            "command": "step-test-targetdisk",
+            "testinfullos": true
+        },
+        {
+            "name": "Test WindowsImage",
+            "command": "step-test-targetwindowsimage",
+            "testinfullos": true
+        },
+        {
+            "name": "Test DriverPack",
+            "command": "step-test-targetdriverpack",
+            "testinfullos": true
+        },
+        {
+            "name": "Remove USB Drive Letters",
+            "command": "step-preinstall-removeusbdriveletter"
+        },
+        {
+            "name": "Clear Local Disk",
+            "command": "step-preinstall-cleartargetdisk"
+        },
+        {
+            "name": "Partition Local Disk",
+            "command": "step-preinstall-partitiontargetdisk"
+        },
+        {
+            "name": "Restore USB Drive Letters",
+            "command": "step-preinstall-restoreusbdriveletter"
+        },
+        {
+            "name": "Enable High Performance Power Plan",
+            "command": "step-preinstall-enablehighperformance"
+        },
+        {
+            "name": "Download Windows ESD from Microsoft",
+            "command": "step-install-downloadwindowsimage",
+            "testinfullos": true
+        },
+        {
+            "name": "Select Windows Image Index",
+            "command": "step-install-getwindowsimageindex",
+            "testinfullos": true
+        },
+        {
+            "name": "Expand Windows Image to Local Disk",
+            "command": "step-install-expandwindowsimage"
+        },
+        {
+            "name": "Restart Logs [C:\\Windows\\Temp\\osdcloud-logs]",
+            "command": "step-install-restartosdcloudlogs",
+            "testinfullos": true
+        },
+        {
+            "name": "Verify Windows Edition on Local Disk",
+            "command": "step-install-getwindowsedition"
+        },
+        {
+            "name": "Remove the downloaded Windows Image",
+            "command": "step-install-removewindowsimage"
+        },
+        {
+            "name": "Apply BCDBoot Configuration",
+            "command": "step-install-bcdboot"
+        },
+        {
+            "name": "Export WinPE Drivers to Local Disk",
+            "command": "step-drivers-recast-winpe"
+        },
+        {
+            "name": "Apply WinPE Drivers to offline Windows Image",
+            "command": "step-drivers-recast-winos"
+        },
+        {
+            "name": "Apply WinPE Drivers to offline Windows Recovery Environment",
+            "command": "step-drivers-recast-winre"
+        },
+        {
+            "name": "Firmware: Download from Microsoft Update Catalog",
+            "command": "step-drivers-firmware",
+            "testinfullos": true
+        },
+        {
+            "name": "Firmware: Apply Driver",
+            "command": "step-drivers-addwindowsdriver-firmware",
+            "skip": false
+        },
+        {
+            "name": "OEM DriverPack: Download from OEM",
+            "command": "step-drivers-driverpack",
+            "testinfullos": true
+        },
+        {
+            "name": "OEM DriverPack: Apply or stage in SetupComplete.cmd",
+            "command": "step-drivers-addwindowsdriver-driverpack"
+        },
+        {
+            "name": "Download Drivers from Microsoft Update Catalog",
+            "command": "step-drivers-msupdate",
+            "testinfullos": true
+        },
+        {
+            "name": "Apply Microsoft Update Drivers - All",
+            "command": "step-drivers-addwindowsdriver-msupdate"
+        },
+        {
+            "name": "Apply Microsoft Update Drivers - Disk",
+            "command": "step-drivers-addwindowsdriver-disk",
+            "skip": true
+        },
+        {
+            "name": "Apply Microsoft Update Drivers - Net",
+            "command": "step-drivers-addwindowsdriver-net"
+        },
+        {
+            "name": "Apply Microsoft Update Drivers - Scsi",
+            "command": "step-drivers-addwindowsdriver-scsi",
+            "skip": true
+        },
+        {
+            "name": "Hotfix for Setup Displayed Eula",
+            "command": "step-update-setupdisplayedeula"
+        },
+        {
+            "name": "Update PowerShell Modules -Offline",
+            "command": "step-powershell-updatemodule"
+        },
+        {
+            "name": "Save PowerShell Module OSD",
+            "command": "step-powershell-savemodule",
+            "parameters": {
+                "name": "OSD"
+            },
+            "skip": true
+        },
+        {
+            "name": "Save PowerShell Module WindowsAutopilotIntune",
+            "command": "step-powershell-savemodule",
+            "parameters": {
+                "name": "WindowsAutopilotIntune"
+            },
+            "skip": true
+        },
+        {
+            "name": "Save PowerShell Module Microsoft.Graph.Groups",
+            "command": "step-powershell-savemodule",
+            "parameters": {
+                "name": "Microsoft.Graph.Groups"
+            },
+            "skip": true
+        },
+        {
+            "name": "Save PowerShell Module Microsoft.Graph.Authentication",
+            "command": "step-powershell-savemodule",
+            "parameters": {
+                "name": "Microsoft.Graph.Authentication"
+            },
+            "skip": true
+        },
+        {
+            "name": "Save PowerShell Module Microsoft.Graph.Identity.DirectoryManagement",
+            "command": "step-powershell-savemodule",
+            "parameters": {
+                "name": "Microsoft.Graph.Identity.DirectoryManagement"
+            },
+            "skip": true
+        },
+        {
+            "name": "Export OS Information to Logs",
+            "command": "step-finalize-exportofflineosinfo"
+        },
+        {
+            "name": "Stop Logs",
+            "command": "step-finalize-osdcloudlogs",
+            "testinfullos": true
+        },
+        {
+            "name": "Stop Workflow",
+            "command": "step-finalize-stoposdcloudworkflow",
+            "testinfullos": true
+        }
+    ]
+}
