+++ workflow/default-back/user-amd64.json
@@ -1,0 +1,29 @@
+{
+    "DriverPacks": {
+        "None": true,
+        "MSCatalog": false,
+        "OSDCloud": true,
+        "Default": "None"
+    },
+    "IsTpmReady": {
+        "Error": "Device is not fully TPM 2.0 compatible.&#x0a;Autopilot will not work.",
+        "Success": "TPM is 2.0 an working as expected.\n Autopilot will work.",
+        "Warning": "Device is not fully TPM 2.0 compatible."
+    },
+    "PSInstallModuleOSD": false,
+    "PSInstallModuleWindowsAutopilotIntune": false,
+    "PSUpdateModulePackageManagement": false,
+    "PSUpdateModulePowershellGet": false,
+    "RecoveryPartition": {
+        "Force": false,
+        "Skip": false,
+        "ToolTip": "Recovery Partition is created by default, except when deploying to a Virtual Machine.&#x0a;Force will always create the Recovery Partition.&#x0a;Skip will never create the Recovery Partition."
+    },
+    "SkipRecoveryPartition": false,
+    "UpdateDiskDrivers": true,
+    "UpdateNetworkDrivers": true,
+    "UpdateScsiDrivers": true,
+    "UpdateSystemFirmware": false,
+    "WinpeRestart": false,
+    "WinpeShutdown": false
+}
