+++ workflow/default-back/ux/OSDCloud.csproj
@@ -1,10 +1,10 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="15.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{EA9F2526-41C0-40AE-B056-BE3A83666ECC}</ProjectGuid>
+    <ProjectGuid>{78491010-EC60-4726-843B-16627447BD36}</ProjectGuid>
     <OutputType>WinExe</OutputType>
     <RootNamespace>OSDCloud</RootNamespace>
     <AssemblyName>OSDCloud</AssemblyName>
