+++ workflow/default-back/ux/OSDCloud.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.11.35312.102
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OSDCloud", "OSDCloud.csproj", "{78491010-EC60-4726-843B-16627447BD36}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{78491010-EC60-4726-843B-16627447BD36}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{78491010-EC60-4726-843B-16627447BD36}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{78491010-EC60-4726-843B-16627447BD36}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{78491010-EC60-4726-843B-16627447BD36}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {2A8E0196-C36D-4B42-81CE-C1603625FA60}
+	EndGlobalSection
+EndGlobal
