+++ workflow/default/ux/MainWindow.xaml.cs
@@ -1,17 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
 
 namespace OSDCloud
 {
@@ -23,11 +11,6 @@ namespace OSDCloud
         public MainWindow()
         {
             InitializeComponent();
-        }
-
-        private void IsAutopilotReady_Click()
-        {
-
         }
     }
 }
