+++ workflow/default/ux/OSDCloud.sln
@@ -1,9 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.11.35312.102
+VisualStudioVersion = 17.13.35931.197 d17.13
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OSDCloud", "OSDCloud.csproj", "{78491010-EC60-4726-843B-16627447BD36}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OSDCloud", "OSDCloud.csproj", "{EA9F2526-41C0-40AE-B056-BE3A83666ECC}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,15 +11,15 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{78491010-EC60-4726-843B-16627447BD36}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{78491010-EC60-4726-843B-16627447BD36}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{78491010-EC60-4726-843B-16627447BD36}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{78491010-EC60-4726-843B-16627447BD36}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EA9F2526-41C0-40AE-B056-BE3A83666ECC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EA9F2526-41C0-40AE-B056-BE3A83666ECC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EA9F2526-41C0-40AE-B056-BE3A83666ECC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EA9F2526-41C0-40AE-B056-BE3A83666ECC}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {2A8E0196-C36D-4B42-81CE-C1603625FA60}
+		SolutionGuid = {3F0FE5B4-04F0-45BF-A820-39F093F5E746}
 	EndGlobalSection
 EndGlobal
