--- workflow/recast/ux/MainWindow.xaml.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Windows;
-
-namespace OSDCloud
-{
-    /// <summary>
-    /// Interaction logic for MainWindow.xaml
-    /// </summary>
-    public partial class MainWindow : Window
-    {
-        public MainWindow()
-        {
-            InitializeComponent();
-        }
-    }
-}
