--- workflow/recast/ux/OSDCloud.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.13.35931.197 d17.13
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OSDCloud", "OSDCloud.csproj", "{EA9F2526-41C0-40AE-B056-BE3A83666ECC}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{EA9F2526-41C0-40AE-B056-BE3A83666ECC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EA9F2526-41C0-40AE-B056-BE3A83666ECC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EA9F2526-41C0-40AE-B056-BE3A83666ECC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EA9F2526-41C0-40AE-B056-BE3A83666ECC}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {3F0FE5B4-04F0-45BF-A820-39F093F5E746}
-	EndGlobalSection
-EndGlobal
